event_management_api/requirements.txt
Django>=3.2.0,<4.0.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=4.8.0
django-filter>=2.4.0
python-dotenv>=0.19.0
django-celery-beat>=2.2.1
google-auth>=2.3.3
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
drf-yasg>=1.20.0